fix calculating the mean in different platforms, see NOTE
calculating the mean of homog/orig stations was different when running on 32 or 64-bit.
it relied on the method numpy.float64.mean() which is not valid for numpy.float.
this conversion to numpy arrays forces to handle nan's manually (using bottleneck.nanmean).

NOTE: because of this, there is a slight difference on the results obtained, which may be
noticed at least on the second decimal place.